feat(namespace): Move namespace routes
