--- app/controllers/project-version/namespace/events.js
@@ -1,1 +1,0 @@
-export { default } from '../../events';
