--- app/controllers/project-version/namespace/index.js
@@ -1,3 +1,0 @@
-import ClassIndexController from "../class/index";
-
-export default ClassIndexController.extend();
