--- app/controllers/project-version/namespace/methods.js
@@ -1,1 +1,0 @@
-export { default } from '../../methods';
