--- app/controllers/project-version/namespace/properties.js
@@ -1,1 +1,0 @@
-export { default } from '../../properties';
