+++ app/controllers/project-version/namespaces/namespace/events.js
@@ -1,0 +1,1 @@
+export { default } from '../../../events';
