+++ app/controllers/project-version/namespaces/namespace/methods.js
@@ -1,0 +1,1 @@
+export { default } from '../../../methods';
