+++ app/controllers/project-version/namespaces/namespace/properties.js
@@ -1,0 +1,1 @@
+export { default } from '../../../properties';
