+++ app/routes/project-version/namespaces/namespace/events.js
