+++ app/routes/project-version/namespaces/namespace/properties.js
